+++ store/in_memory_player_store.go
@@ -1,27 +1,59 @@
 package store
 
 import (
-	"errors"
+	"sort"
+	"sync"
 )
 
-// InMemoryPlayerStore represents a PlayerStore that is stored in
-// memory. In a future release, this will be replaced with a
-// persistent store
+// InMemoryPlayerStore represents a PlayerStore that is stored in memory
 type InMemoryPlayerStore struct {
+	mu    sync.Mutex
 	store map[string]int
 }
 
+// NewInMemoryPlayerStore returns an InMemoryPlayerStore with an empty store
 func NewInMemoryPlayerStore() *InMemoryPlayerStore {
-	return &InMemoryPlayerStore{map[string]int{}}
+	i := new(InMemoryPlayerStore)
+	i.store = map[string]int{}
+	return i
 }
 
+// GetPlayerScore returns a score if the player exists or an error if
+// they don't
 func (i *InMemoryPlayerStore) GetPlayerScore(name string) (int, error) {
+	i.mu.Lock()
+	defer i.mu.Unlock()
 	if _, ok := i.store[name]; !ok {
-		return 0, errors.New("player not found")
+		return 0, ErrPlayerNotFound
 	}
 	return i.store[name], nil
 }
 
+// RecordWin increments a player's score by 1. If the player doesn't exist,
+// it will create the player and increment their score to 1.
 func (i *InMemoryPlayerStore) RecordWin(name string) {
+	i.mu.Lock()
+	defer i.mu.Unlock()
 	i.store[name]++
+}
+
+// GetLeague returns an ordered slice containing every Player in the league
+// sorted by score, descending
+func (i *InMemoryPlayerStore) GetLeague() []Player {
+	league := i.getUnsortedLeague()
+
+	sort.Sort(sort.Reverse(league))
+	return league
+}
+
+func (i *InMemoryPlayerStore) getUnsortedLeague() PlayerSlice {
+	league := make(PlayerSlice, len(i.store))
+	idx := 0
+	i.mu.Lock()
+	defer i.mu.Unlock()
+	for k, v := range i.store {
+		league[idx] = Player{Name: k, Wins: v}
+		idx++
+	}
+	return league
 }