+++ store/in_memory_player_store_test.go
@@ -1,55 +1,44 @@
 package store_test
 
 import (
-	"errors"
+	"reflect"
 	"testing"
 
 	"github.com/nelsen129/player-league/store"
 )
 
-// InMemoryPlayerStore represents a PlayerStore that is stored in
-// memory. In a future release, this will be replaced with a
-// persistent store
-type InMemoryPlayerStore struct {
-	store map[string]int
-}
-
-func NewInMemoryPlayerStore() *InMemoryPlayerStore {
-	return &InMemoryPlayerStore{map[string]int{}}
-}
-
-func (i *InMemoryPlayerStore) GetPlayerScore(name string) (int, error) {
-	if _, ok := i.store[name]; !ok {
-		return 0, errors.New("player not found")
-	}
-	return i.store[name], nil
-}
-
-func (i *InMemoryPlayerStore) RecordWin(name string) {
-	i.store[name]++
-}
-
 func TestInMemoryPlayerStore(t *testing.T) {
-	t.Run("returns an error on get if player doesn't exist", func(t *testing.T) {
+	t.Run("standard player store suite", func(t *testing.T) {
 		playerStore := store.NewInMemoryPlayerStore()
-		_, err := playerStore.GetPlayerScore("Bob")
-		if err == nil {
-			t.Error("want error, got none")
-		}
+		testStore(t, playerStore)
 	})
 
-	t.Run("records and returns a score for a new player", func(t *testing.T) {
+	t.Run("sorts for GetLeague", func(t *testing.T) {
 		playerStore := store.NewInMemoryPlayerStore()
-		playerStore.RecordWin("Neil")
-		got, err := playerStore.GetPlayerScore("Neil")
-		want := 1
-
-		if err != nil {
-			t.Errorf("want no error, got %v", err)
+		sortedLeague := []store.Player{
+			{Name: "Alice", Wins: 5},
+			{Name: "Bob", Wins: 4},
+			{Name: "Charlie", Wins: 3},
+			{Name: "Dave", Wins: 2},
+			{Name: "Eve", Wins: 1},
 		}
-		
-		if got != want {
-			t.Errorf("got %d, want %d", got, want)
+		unsortedLeague := []store.Player{
+			{Name: "Dave", Wins: 2},
+			{Name: "Bob", Wins: 4},
+			{Name: "Charlie", Wins: 3},
+			{Name: "Eve", Wins: 1},
+			{Name: "Alice", Wins: 5},
+		}
+
+		for _, player := range unsortedLeague {
+			for range player.Wins {
+				playerStore.RecordWin(player.Name)
+			}
+		}
+
+		got := playerStore.GetLeague()
+		if !reflect.DeepEqual(got, sortedLeague) {
+			t.Errorf("got %v, want %v", got, sortedLeague)
 		}
 	})
 }